requirements.txt
django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
gunicorn==20.1.0
django-heroku==0.3.1
tweepy==4.10.1
python-decouple==3.6
djongo==1.3.6
APScheduler==3.9.1
